Refactor bff file structure
