+++ src/app/(routes)/(noAuth)/signup/signup-form.spec.tsx
@@ -1,7 +1,7 @@
 import { render, screen } from '@testing-library/react';
 import '@testing-library/jest-dom';
 import SignupForm from './signup-form';
-import { SignUpDto, signup } from '@/app/bff-utils/auth-utils';
+import { SignUpDto, signup } from '@/app/bff-utils/auth/auth-utils';
 import routes from '@/common/routes';
 import userEvent from '@testing-library/user-event';
 
@@ -25,7 +25,7 @@ jest.mock('../../../../common/forms/use-form-validation', () => ({
   })
 }));
 
-jest.mock('../../../bff-utils/auth-utils', () => ({
+jest.mock('../../../bff-utils/auth/auth-utils', () => ({
   signup: jest.fn()
 }));
